add @flask_endpoint to log and debug endpoints
this decorator lets you debug a flask endpoint using statsd
and logging.

the statsd side of the endpoint decorator gives you
two important factors, number of times the endpoint was hit and
the time it took to exec the endpoint. statsd uses udp, so there's
no infrastructure dependencies required.

the logging side constructs a log message using the flask request
and then writes it to info.logger

you can tune what % of requests are logged through the log_percent var,
and the is_active param lets you disable the decorator regardless of the
log_percent